Reorganise events and event writers
