+++ simulator/opendc-experiments/opendc-experiments-allocateam/src/main/kotlin/org/opendc/experiments/allocateam/telemetry/events/JobFinishedEvent.kt
@@ -1,4 +1,4 @@
-package org.opendc.experiments.allocateam.telemetry
+package org.opendc.experiments.allocateam.telemetry.events
 
 import org.opendc.experiments.sc20.telemetry.Event
